# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News API",
        "description": "API для управління новинами з блоковою структурою контенту.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/news": {
            "get": {
                "tags": [
                    "Public News"
                ],
                "summary": "Отримати список опублікованих новин",
                "operationId": "c33e145d81ca885577167b753733eca0",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Пошук по назві або опису",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Кількість на сторінку (макс 50)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер сторінки",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список новин"
                    }
                }
            }
        },
        "/api/news/{slug}": {
            "get": {
                "tags": [
                    "Public News"
                ],
                "summary": "Отримати новину за slug (з блоками контенту)",
                "operationId": "79799729411d70e82529ca8210936edc",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug новини",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Дані новини з блоками"
                    },
                    "404": {
                        "description": "Новину не знайдено"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Введіть токен, отриманий при логіні",
                "bearerFormat": "sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}